requirements.txt
flask==3.0.2
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
elevenlabs==0.2.21
boto3==1.34.1
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg==3.1.17
psycopg-pool==3.2.1
flask-migrate==4.0.5
flask-admin==1.6.1
wtforms==3.1.2
pyjwt==2.8.0
resend==0.6.0
email-validator==2.1.0.post1
redis==5.0.1
celery==5.3.6
cartesia==1.0.0